+++ example/android/app/src/main/kotlin/com/pinzarulab/smart_domain_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.pinzarulab.smart_domain_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()